+++ espnet/bin/asr_recog.py
@@ -4,30 +4,42 @@
 # Copyright 2017 Johns Hopkins University (Shinji Watanabe)
 #  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
+
+import argparse
 import logging
 import os
+import random
 import sys
 
-from espnet.bin.bin_utils import check_cuda_visible_devices
-from espnet.bin.bin_utils import get_recog_argparser
-from espnet.bin.bin_utils import set_logging_level
-from espnet.bin.bin_utils import set_seed
+import numpy as np
 
 
 def main(args):
-    parser = get_recog_argparser('asr')
+    parser = argparse.ArgumentParser()
     # general configuration
+    parser.add_argument('--ngpu', default=0, type=int,
+                        help='Number of GPUs')
+    parser.add_argument('--backend', default='chainer', type=str,
+                        choices=['chainer', 'pytorch'],
+                        help='Backend library')
     parser.add_argument('--debugmode', default=1, type=int,
                         help='Debugmode')
     parser.add_argument('--seed', default=1, type=int,
                         help='Random seed')
-    parser.add_argument('--batch-size', default=1, type=int,
+    parser.add_argument('--verbose', '-V', default=1, type=int,
+                        help='Verbose option')
+    parser.add_argument('--batchsize', default=1, type=int,
                         help='Batch size for beam search (0: means no batch processing)')
     # task related
     parser.add_argument('--recog-json', type=str,
                         help='Filename of recognition data (json)')
     parser.add_argument('--result-label', type=str, required=True,
                         help='Filename of result label data (json)')
+    # model (parameter) related
+    parser.add_argument('--model', type=str, required=True,
+                        help='Model file parameters to read')
+    parser.add_argument('--model-conf', type=str, default=None,
+                        help='Model config file')
     # search related
     parser.add_argument('--nbest', type=int, default=1,
                         help='Output N-best hypotheses')
@@ -35,6 +47,12 @@ def main(args):
                         help='Beam size')
     parser.add_argument('--penalty', default=0.0, type=float,
                         help='Incertion penalty')
+    parser.add_argument('--maxlenratio', default=0.0, type=float,
+                        help="""Input length ratio to obtain max output length.
+                        If maxlenratio=0.0 (default), it uses a end-detect function
+                        to automatically find maximum hypothesis lengths""")
+    parser.add_argument('--minlenratio', default=0.0, type=float,
+                        help='Input length ratio to obtain min output length')
     parser.add_argument('--ctc-weight', default=0.0, type=float,
                         help='CTC weight in joint decoding')
     # rnnlm related
@@ -52,15 +70,39 @@ def main(args):
                         help='RNNLM weight.')
     args = parser.parse_args(args)
 
-    set_logging_level(args.verbose)
+    # logging info
+    if args.verbose == 1:
+        logging.basicConfig(
+            level=logging.INFO, format="%(asctime)s (%(module)s:%(lineno)d) %(levelname)s: %(message)s")
+    elif args.verbose == 2:
+        logging.basicConfig(level=logging.DEBUG,
+                            format="%(asctime)s (%(module)s:%(lineno)d) %(levelname)s: %(message)s")
+    else:
+        logging.basicConfig(
+            level=logging.WARN, format="%(asctime)s (%(module)s:%(lineno)d) %(levelname)s: %(message)s")
+        logging.warning("Skip DEBUG/INFO messages")
 
-    # TODO(mn5k): support of multiple GPUs
-    check_cuda_visible_devices(args.ngpu, 1)
+    # check CUDA_VISIBLE_DEVICES
+    if args.ngpu > 0:
+        cvd = os.environ.get("CUDA_VISIBLE_DEVICES")
+        if cvd is None:
+            logging.warning("CUDA_VISIBLE_DEVICES is not set.")
+        elif args.ngpu != len(cvd.split(",")):
+            logging.error("#gpus is not matched with CUDA_VISIBLE_DEVICES.")
+            sys.exit(1)
+
+        # TODO(mn5k): support of multiple GPUs
+        if args.ngpu > 1:
+            logging.error("The program only supports ngpu=1.")
+            sys.exit(1)
 
     # display PYTHONPATH
     logging.info('python path = ' + os.environ.get('PYTHONPATH', '(None)'))
 
-    set_seed(args.seed)
+    # seed setting
+    random.seed(args.seed)
+    np.random.seed(args.seed)
+    logging.info('set random seed = %d' % args.seed)
 
     # recog
     logging.info('backend = ' + args.backend)
