+++ espnet/bin/asr_train.py
@@ -3,25 +3,43 @@
 # Copyright 2017 Tomoki Hayashi (Nagoya University)
 #  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
+
+import argparse
 import logging
 import os
+import platform
+import random
+import subprocess
 import sys
 
-from espnet.bin.bin_utils import check_cuda_visible_devices
-from espnet.bin.bin_utils import get_train_argparser
-from espnet.bin.bin_utils import set_logging_level
-from espnet.bin.bin_utils import set_seed
+import numpy as np
 
 
 def main(args):
-    parser = get_train_argparser('asr')
+    parser = argparse.ArgumentParser()
     # general configuration
+    parser.add_argument('--ngpu', default=0, type=int,
+                        help='Number of GPUs')
+    parser.add_argument('--backend', default='chainer', type=str,
+                        choices=['chainer', 'pytorch'],
+                        help='Backend library')
+    parser.add_argument('--outdir', type=str, required=True,
+                        help='Output directory')
+    parser.add_argument('--debugmode', default=1, type=int,
+                        help='Debugmode')
     parser.add_argument('--dict', required=True,
                         help='Dictionary')
+    parser.add_argument('--seed', default=1, type=int,
+                        help='Random seed')
     parser.add_argument('--debugdir', type=str,
                         help='Output directory for debugging')
+    parser.add_argument('--resume', '-r', default='', nargs='?',
+                        help='Resume the training from snapshot')
     parser.add_argument('--minibatches', '-N', type=int, default='-1',
                         help='Process only N minibatches (for debug)')
+    parser.add_argument('--verbose', '-V', default=0, type=int,
+                        help='Verbose option')
+    parser.add_argument('--tensorboard-dir', default=None, type=str, nargs='?', help="Tensorboard log dir path")
     # task related
     parser.add_argument('--train-json', type=str, default=None,
                         help='Filename of train label data (json)')
@@ -113,6 +131,8 @@ def main(args):
     parser.add_argument('--dropout-rate', default=0.0, type=float,
                         help='Dropout rate')
     # minibatch related
+    parser.add_argument('--batch-size', '-b', default=50, type=int,
+                        help='Batch size')
     parser.add_argument('--maxlen-in', default=800, type=int, metavar='ML',
                         help='Batch size is reduced if the input sequence length > ML')
     parser.add_argument('--maxlen-out', default=150, type=int, metavar='ML',
@@ -132,18 +152,55 @@ def main(args):
                         help='Criterion to perform epsilon decay')
     parser.add_argument('--threshold', default=1e-4, type=float,
                         help='Threshold to stop iteration')
+    parser.add_argument('--epochs', '-e', default=30, type=int,
+                        help='Maximum number of epochs')
+    parser.add_argument('--early-stop-criterion', default='validation/main/loss', type=str, nargs='?',
+                        help="Value to monitor to trigger an early stopping of the training")
+    parser.add_argument('--patience', default=0, type=int, nargs='?',
+                        help="Number of epochs to wait without improvement before stopping the training")
+    parser.add_argument('--grad-clip', default=5, type=float,
+                        help='Gradient norm threshold to clip')
     parser.add_argument('--num-save-attention', default=3, type=int,
                         help='Number of samples of attention to be saved')
     args = parser.parse_args(args)
 
-    set_logging_level(args.verbose)
+    # logging info
+    if args.verbose > 0:
+        logging.basicConfig(
+            level=logging.INFO, format='%(asctime)s (%(module)s:%(lineno)d) %(levelname)s: %(message)s')
+    else:
+        logging.basicConfig(
+            level=logging.WARN, format='%(asctime)s (%(module)s:%(lineno)d) %(levelname)s: %(message)s')
+        logging.warning('Skip DEBUG/INFO messages')
 
-    check_cuda_visible_devices(args.ngpu)
+    # check CUDA_VISIBLE_DEVICES
+    if args.ngpu > 0:
+        # python 2 case
+        if platform.python_version_tuple()[0] == '2':
+            if "clsp.jhu.edu" in subprocess.check_output(["hostname", "-f"]):
+                cvd = subprocess.check_output(["/usr/local/bin/free-gpu", "-n", str(args.ngpu)]).strip()
+                logging.info('CLSP: use gpu' + cvd)
+                os.environ['CUDA_VISIBLE_DEVICES'] = cvd
+        # python 3 case
+        else:
+            if "clsp.jhu.edu" in subprocess.check_output(["hostname", "-f"]).decode():
+                cvd = subprocess.check_output(["/usr/local/bin/free-gpu", "-n", str(args.ngpu)]).decode().strip()
+                logging.info('CLSP: use gpu' + cvd)
+                os.environ['CUDA_VISIBLE_DEVICES'] = cvd
+        cvd = os.environ.get("CUDA_VISIBLE_DEVICES")
+        if cvd is None:
+            logging.warning("CUDA_VISIBLE_DEVICES is not set.")
+        elif args.ngpu != len(cvd.split(",")):
+            logging.error("#gpus is not matched with CUDA_VISIBLE_DEVICES.")
+            sys.exit(1)
 
     # display PYTHONPATH
     logging.info('python path = ' + os.environ.get('PYTHONPATH', '(None)'))
 
-    set_seed(args.seed)
+    # set random seed
+    logging.info('random seed = %d' % args.seed)
+    random.seed(args.seed)
+    np.random.seed(args.seed)
 
     # load dictionary for debug log
     if args.dict is not None:
