+++ espnet/bin/lm_train.py
@@ -9,21 +9,37 @@
 from __future__ import division
 from __future__ import print_function
 
+import argparse
 import logging
+import numpy as np
 import os
+import platform
+import random
+import subprocess
 import sys
-
-from espnet.bin.bin_utils import check_cuda_visible_devices
-from espnet.bin.bin_utils import get_train_argparser
-from espnet.bin.bin_utils import set_logging_level
-from espnet.bin.bin_utils import set_seed
 
 
 def main(args):
-    parser = get_train_argparser('lm')
+    parser = argparse.ArgumentParser()
     # general configuration
+    parser.add_argument('--ngpu', default=0, type=int,
+                        help='Number of GPUs')
+    parser.add_argument('--backend', default='chainer', type=str,
+                        choices=['chainer', 'pytorch'],
+                        help='Backend library')
+    parser.add_argument('--outdir', type=str, required=True,
+                        help='Output directory')
+    parser.add_argument('--debugmode', default=1, type=int,
+                        help='Debugmode')
     parser.add_argument('--dict', type=str, required=True,
                         help='Dictionary')
+    parser.add_argument('--seed', default=1, type=int,
+                        help='Random seed')
+    parser.add_argument('--resume', '-r', default='', nargs='?',
+                        help='Resume the training from snapshot')
+    parser.add_argument('--verbose', '-V', default=0, type=int,
+                        help='Verbose option')
+    parser.add_argument('--tensorboard-dir', default=None, type=str, nargs='?', help="Tensorboard log dir path")
     # task related
     parser.add_argument('--train-label', type=str, required=True,
                         help='Filename of train label data')
@@ -35,6 +51,16 @@ def main(args):
     parser.add_argument('--opt', default='sgd', type=str,
                         choices=['sgd', 'adam'],
                         help='Optimizer')
+    parser.add_argument('--batchsize', '-b', type=int, default=300,
+                        help='Number of examples in each mini-batch')
+    parser.add_argument('--epoch', '-e', type=int, default=20,
+                        help='Number of sweeps over the dataset to train')
+    parser.add_argument('--early-stop-criterion', default='validation/main/loss', type=str, nargs='?',
+                        help="Value to monitor to trigger an early stopping of the training")
+    parser.add_argument('--patience', default=0, type=int, nargs='?',
+                        help="Number of epochs to wait without improvement before stopping the training")
+    parser.add_argument('--gradclip', '-c', type=float, default=5,
+                        help='Gradient norm threshold to clip')
     parser.add_argument('--layer', '-l', type=int, default=2,
                         help='Number of hidden layers')
     parser.add_argument('--unit', '-u', type=int, default=650,
@@ -43,14 +69,43 @@ def main(args):
                         help='Batch size is reduced if the input sequence > ML')
     args = parser.parse_args(args)
 
-    set_logging_level(args.verbose)
+    # logging info
+    if args.verbose > 0:
+        logging.basicConfig(
+            level=logging.INFO, format='%(asctime)s (%(module)s:%(lineno)d) %(levelname)s: %(message)s')
+    else:
+        logging.basicConfig(
+            level=logging.WARN, format='%(asctime)s (%(module)s:%(lineno)d) %(levelname)s: %(message)s')
+        logging.warning('Skip DEBUG/INFO messages')
 
-    check_cuda_visible_devices(args.ngpu)
+    # check CUDA_VISIBLE_DEVICES
+    if args.ngpu > 0:
+        # python 2 case
+        if platform.python_version_tuple()[0] == '2':
+            if "clsp.jhu.edu" in subprocess.check_output(["hostname", "-f"]):
+                cvd = subprocess.check_output(["/usr/local/bin/free-gpu", "-n", str(args.ngpu)]).strip()
+                logging.info('CLSP: use gpu' + cvd)
+                os.environ['CUDA_VISIBLE_DEVICES'] = cvd
+        # python 3 case
+        else:
+            if "clsp.jhu.edu" in subprocess.check_output(["hostname", "-f"]).decode():
+                cvd = subprocess.check_output(["/usr/local/bin/free-gpu", "-n", str(args.ngpu)]).decode().strip()
+                logging.info('CLSP: use gpu' + cvd)
+                os.environ['CUDA_VISIBLE_DEVICES'] = cvd
+        cvd = os.environ.get("CUDA_VISIBLE_DEVICES")
+        if cvd is None:
+            logging.warning("CUDA_VISIBLE_DEVICES is not set.")
+        elif args.ngpu != len(cvd.split(",")):
+            logging.error("#gpus is not matched with CUDA_VISIBLE_DEVICES.")
+            sys.exit(1)
 
     # display PYTHONPATH
     logging.info('python path = ' + os.environ.get('PYTHONPATH', '(None)'))
 
-    set_seed(args.seed)
+    # seed setting
+    nseed = args.seed
+    random.seed(nseed)
+    np.random.seed(nseed)
 
     # load dictionary
     with open(args.dict, 'rb') as f:
