--- espnet/utils/chainer_utils.py
@@ -1,13 +1,0 @@
-import chainer
-
-
-def chainer_load(path, model):
-    """Function to load chainer model parameters
-
-    :param str path: model file or snapshot file to be loaded
-    :param chainer.Chain model: chainer model
-    """
-    if 'snapshot' in path:
-        chainer.serializers.load_npz(path, model, path='updater/model:main/')
-    else:
-        chainer.serializers.load_npz(path, model)
